requirements.txt
pocketsphinx
grpcio>=1.1.0
PyAudio>=0.2.10
proto-google-cloud-speech-v1beta1>=0.15.1
six>=1.10.0
requests>=2.13.0
google-auth>=0.8.0
numpy
PyWavelets
PyYaml